renamed lib to pkg. moved main to cmd/cli.
