+++ cmd/cli/main_test.go
@@ -1,0 +1,23 @@
+package main
+
+import (
+	"testing"
+)
+
+func Test_self(t *testing.T) {
+	tests := []struct {
+		name    string
+		wantErr bool
+	}{
+		{"expected", false},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			_, err := self()
+			if (err != nil) != tt.wantErr {
+				t.Errorf("self() error = %v, wantErr %v", err, tt.wantErr)
+				return
+			}
+		})
+	}
+}
