--- lib/geolite2/db/GeoLite2-City/README.txt
@@ -1,1 +1,0 @@
-Latitude and longitude are not precise and should not be used to identify a particular street address or household.
