--- lib/geolite2/db/GeoLite2-Country/COPYRIGHT.txt
@@ -1,1 +1,0 @@
-Database and Contents Copyright (c) 2021 MaxMind, Inc.
