+++ pkg/geolite2/db/GeoLite2-Country/COPYRIGHT.txt
@@ -1,0 +1,1 @@
+Database and Contents Copyright (c) 2021 MaxMind, Inc.
